+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## Version 0.1 (development)
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Add query
+- Add scripts
+- Setup Poetry